Add files via upload
This scripts used to preparing egohand dataset and  train a hand detector using coco pre-trained weights for transfer learning. As with any DNN based task, the most expensive (and riskiest) part of the process has to do with finding or creating the right (annotated) dataset. I was interested mainly in detecting hands on a table (egocentric view point).  I Experiment with Egohands Dataset which was a much better fit to my requirements.